[Feat] #30 - FfipCameraHeaderBar 추가
